requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
pmdarima==2.0.4
statsmodels==0.14.1
joblib==1.3.2
python-multipart==0.0.6
pyserial==3.5
websockets==15.0.1
requests==2.31.0
aiohttp==3.9.1
